+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 ## [Unreleased]
 
+- Rename `Obsidian::MarkdownConverter` to `Obsidian::ParsedMarkdownDocument`
+
 ## [0.3.0] - 2023-07-27
 
-- Note objects have a `parent` attribute.
+- `Note` objects have a `parent` attribute.
 
 ## [0.2.0] - 2023-07-24
 
-- Note objects have a `content` attribute. Call `content.generate_html` to generate HTML on demand.
+- `Note` objects have a `content` attribute. Call `content.generate_html` to generate HTML on demand.
 
 ## [0.1.0] - 2023-07-10