+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased]
 
 - Rename `Obsidian::MarkdownConverter` to `Obsidian::ParsedMarkdownDocument`
-- Unify `Note` and `Index` classes into `Page`. This is a breaking API change.
+- Unify `Note` and `Index` classes into `Page`. This is a breaking API change. `Parser#notes is replaced by Parse#pages`. Call `Page#is_index?`to distinguish between directory derived pages and documents.
+- Removed `Parser#table_of_contents`
 
 ## [0.3.0] - 2023-07-27